Refactor db connection logic
